+++ syntax/attribute_value.ml
@@ -582,6 +582,8 @@ let sandbox = spaces variant
 
 let popover = variant_or_empty "Auto"
 
+let command = total_variant Html_types_reflected.command_value
+
 let in_ = total_variant Svg_types_reflected.in_value
 
 let in2 = in_